+++ leviculum-std/src/interfaces/byte_channel/mod.rs
@@ -82,6 +82,7 @@ where
             hw_mtu: Some(BYTE_CHANNEL_HW_MTU),
             is_local_client: false,
             bitrate: None,
+            announce_cap_bitrate: None,
             tx_jitter_max_ms: None,
             ifac: None,
             mode: leviculum_core::traits::InterfaceMode::default(),